+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 - Fix bug where gated value icons didn't respect Hide filters.
 - Change random objective selection UI to support type-to-filter (at the cost
   of losing the subheadings in the dropdown list).
-- Add colour pickers for the background and text colours.
+- Add colour pickers for the background and text colours, which persist between
+  pageloads.
 
 ## 2021-08-18
 - Mark objectives as complete when their locations are dismissed, since that's